Further Division Work
Adding a more enhanced method now of detecting when a cell should
divide. Comparing the cells time "alive" and dividing when it reaches
that point. Need to touch up on it a lot but the groundwork is there. I
will add some variance in future when the code works for the meantime.