Cell Size Confirmation
Added a function in my Root_Model.py code that would print the length of
the cell to a Data file which I could use to produce a Plot of cell
length.

Noticed error with my division rate so have altered that and appended
the correction to the Root_Model.py script.